requirement.txt
python-docx >= 0.8.10
pandas >= 1.2
openpyxl >= 3.0.6
lxml >= 2.3.2